Create a short, complete explanation of req. info.
I don't need to have nearly so much information explaining how to use
optional command line arguments. The quirk about the "repo" needing to
be a URL fragment somewhat justified the extra explanation, but that's
gone now.

Instead, a short, up-front section stating which bits are required and
where the program will try to get them.